# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger документация проекта <<В добрые руки>>",
        "description": "Open API",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://in-good-hands.dev.mind4.me",
            "description": " API Server"
        }
    ],
    "paths": {
...
        "/api/create_review": {
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Создание Отзыва",
                "operationId": "92ca9a4678598e9f620af70b52eb7d95",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_user_owner",
                                    "score",
                                    "text"
                                ],
                                "properties": {
                                    "id_reservation": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "score": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Новый отзыв"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/get_user_reviews": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Посмотреть отзывы пользователя",
                "operationId": "601331b86cc70ef33604c04b56ba699b",
                "parameters": [
                    {
                        "name": "id_user_owner",
                        "in": "query",
                        "required": true,
                        "example": 9
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "example": 3
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "required": false,
                        "example": "4"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/get_my_reviews": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Посмотреть отзывы на себя",
                "operationId": "6c2371fc8b95268ff49c93a74e058dff",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "example": 3
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "required": false,
                        "example": "3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/api/get_short_user_info": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Получить краткую информацию о пользователе",
                "operationId": "3223c2e1573a97e52b262e68fa4226a7",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "query",
                        "required": true,
                        "example": 8
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/change_user_info": {
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Изменение данных пользователя",
                "operationId": "dc064da01c05e88bfd68fdd4707ff31c",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Новое имя"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "new@.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+7(123)4567890"
                                    },
                                    "id_city": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "402": {
                        "description": "Incorrect code"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://in-good-hands.dev.mind4.me/oauth/authorize",
                        "tokenUrl": "https://in-good-hands.dev.mind4.me/oauth/token",
                        "refreshUrl": "https://in-good-hands.dev.mind4.me/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}